refactor(Utils) Cleanup Utils class, improve separation of concerns (SoC)
