+++ src/main/java/org/cloudfoundry/community/servicebroker/postgresql/service/Database.java
@@ -21,7 +21,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
-import java.sql.SQLException;
+import java.sql.*;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -33,8 +33,8 @@ public class Database {
 
     public void createDatabaseForInstance(String instanceId, String serviceId, String planId, String organizationGuid, String spaceGuid) throws SQLException {
         Utils.checkValidUUID(instanceId);
-        Utils.executeUpdate("CREATE DATABASE \"" + instanceId + "\" ENCODING 'UTF8'");
-        Utils.executeUpdate("REVOKE all on database \"" + instanceId + "\" from public");
+        PostgreSQLDatabase.executeUpdate("CREATE DATABASE \"" + instanceId + "\" ENCODING 'UTF8'");
+        PostgreSQLDatabase.executeUpdate("REVOKE all on database \"" + instanceId + "\" from public");
 
         Map<Integer, String> parameterMap = new HashMap<Integer, String>();
         parameterMap.put(1, instanceId);
@@ -43,7 +43,7 @@ public class Database {
         parameterMap.put(4, organizationGuid);
         parameterMap.put(5, spaceGuid);
 
-        Utils.executePreparedUpdate("INSERT INTO service (serviceinstanceid, servicedefinitionid, planid, organizationguid, spaceguid) VALUES (?, ?, ?, ?, ?)", parameterMap);
+        PostgreSQLDatabase.executePreparedUpdate("INSERT INTO service (serviceinstanceid, servicedefinitionid, planid, organizationguid, spaceguid) VALUES (?, ?, ?, ?, ?)", parameterMap);
     }
 
     public void deleteDatabase(String instanceId) throws SQLException {
@@ -53,7 +53,7 @@ public class Database {
         parameterMap.put(1, instanceId);
 
         try {
-            Map<String, String> result = Utils.executeSelect("SELECT current_user");
+            Map<String, String> result = PostgreSQLDatabase.executeSelect("SELECT current_user");
             String currentUser = null;
 
             if(result != null) {
@@ -64,10 +64,10 @@ public class Database {
                 logger.error("Current user could not be found?");
             }
 
-            Utils.executePreparedSelect("SELECT pg_terminate_backend(pg_stat_activity.pid) FROM pg_stat_activity WHERE pg_stat_activity.datname = ? AND pid <> pg_backend_pid()", parameterMap);
-            Utils.executeUpdate("ALTER DATABASE \"" + instanceId + "\" OWNER TO \"" + currentUser + "\"");
-            Utils.executeUpdate("DROP DATABASE IF EXISTS \"" + instanceId + "\"");
-            Utils.executePreparedUpdate("DELETE FROM service WHERE serviceinstanceid=?", parameterMap);
+            PostgreSQLDatabase.executePreparedSelect("SELECT pg_terminate_backend(pg_stat_activity.pid) FROM pg_stat_activity WHERE pg_stat_activity.datname = ? AND pid <> pg_backend_pid()", parameterMap);
+            PostgreSQLDatabase.executeUpdate("ALTER DATABASE \"" + instanceId + "\" OWNER TO \"" + currentUser + "\"");
+            PostgreSQLDatabase.executeUpdate("DROP DATABASE IF EXISTS \"" + instanceId + "\"");
+            PostgreSQLDatabase.executePreparedUpdate("DELETE FROM service WHERE serviceinstanceid=?", parameterMap);
         } catch (SQLException e) {
             logger.error(e.getMessage());
         }
@@ -79,7 +79,7 @@ public class Database {
         Map<Integer, String> parameterMap = new HashMap<Integer, String>();
         parameterMap.put(1, instanceId);
 
-        Map<String, String> result = Utils.executePreparedSelect("SELECT * FROM service WHERE serviceinstanceid = ?", parameterMap);
+        Map<String, String> result = PostgreSQLDatabase.executePreparedSelect("SELECT * FROM service WHERE serviceinstanceid = ?", parameterMap);
 
         String serviceDefinitionId = result.get("servicedefinitionid");
         String organizationGuid = result.get("organizationguid");
